Better tokenizer.

+++ src/com/sciolizer/jbscript/lang/ast/statement/Gosub.java
@@ -3,11 +3,13 @@ package com.sciolizer.jbscript.lang.ast.statement;
 import com.sciolizer.jbscript.lang.ast.Statement;
 import com.sciolizer.jbscript.lang.ast.StatementVisitor;
 
+import java.math.BigInteger;
+
 // First created by jball on 8/23/13 at 10:35 PM
 public class Gosub implements Statement {
-    public final int lineNumber;
+    public final BigInteger lineNumber;
 
-    public Gosub(int lineNumber) {
+    public Gosub(BigInteger lineNumber) {
         this.lineNumber = lineNumber;
     }
