+++ src/com/sciolizer/jbscript/lang/ast/statement/Goto.java
@@ -3,11 +3,13 @@ package com.sciolizer.jbscript.lang.ast.statement;
 import com.sciolizer.jbscript.lang.ast.Statement;
 import com.sciolizer.jbscript.lang.ast.StatementVisitor;
 
+import java.math.BigInteger;
+
 // First created by jball on 8/23/13 at 10:33 PM
 public class Goto implements Statement {
-    public final int lineNumber;
+    public final BigInteger lineNumber;
 
-    public Goto(int lineNumber) {
+    public Goto(BigInteger lineNumber) {
         this.lineNumber = lineNumber;
     }
